Chart corrected

+++ components/therapyAnalysis/src/specificworker.h
@@ -141,7 +141,7 @@ private:
 	float getElbowAngleVec(std::string side);
 	float getAngleBetweenVectors(QVec v1, QVec v2);
 	float getDeviation(std::string part);
-	void saveActualFrameMetrics(long int time);
+	void saveActualFrameMetrics(float time);
 	void initializeMetrics();
 //	void calculateAllMetrics();
